rename: 대시보드 서브페이지 파일 구조화(#99)
