+++ src/pages/dashboard/sub-pages/Reason/Reason.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react'
-import Section from './Section'
-import { DROPDOWN_REASON_OPTIONS } from './constants'
-import type { DropdownProps } from '@/components/ui/Dropdown/Dropdown.types'
-import DoughnutChart from '../../../components/charts/DoughnutChart/DoughnutChart'
-import VerticalBarChart from '../../../components/charts/VerticalBarChart/VerticalBarChart'
+import Section from '../Section'
+import { DROPDOWN_REASON_OPTIONS } from '../constants'
+import type { DropdownProps } from '@components/ui/Dropdown/Dropdown.types'
+import DoughnutChart from '@components/charts/DoughnutChart/DoughnutChart'
+import VerticalBarChart from '@components/charts/VerticalBarChart/VerticalBarChart'
 
 export default function Reason() {
   const [reason, setReason] = useState('DISSATISFACTION')
